requirements.txt
Flask
Selenium
pymongo
python-dotenv
requests
beautifulsoup4
webdriver-manager